requirements.txt
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
einops==0.7.0
filelock==3.13.1
fsspec==2024.3.1
gmpy2==2.1.2
huggingface-hub==0.21.4
idna==3.4
imageio==2.34.0
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pip==23.3.1
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.
PySocks==1.7.1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
setuptools==68.2.2
sympy==1.12
tifffile==2024.2.12
timm==0.9.16
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
wheel==0.41.2
